@@ -6,7 +6,7 @@ author = Robert T. McGibbon
 author-email = rmcgibbo@gmail.com
 summary = Cross-distribution Linux wheels
 description-file = README.rst
-python_requires = >=3.5
+python_requires = >=3.6
 classifier =
         Development Status :: 4 - Beta
         Environment :: Console
@@ -14,7 +14,6 @@ classifier =
         License :: OSI Approved :: MIT License
         Operating System :: POSIX :: Linux
         Programming Language :: Python :: 3
-        Programming Language :: Python :: 3.5
         Programming Language :: Python :: 3.6
         Programming Language :: Python :: 3.7
         Programming Language :: Python :: 3.8
