@@ -24,15 +24,15 @@ def test_patchelf_check_output_fail(check_output):
 @patch("auditwheel.patcher.check_output")
 @pytest.mark.parametrize("version", ["0.9", "0.9.1", "0.10"])
 def test_patchelf_version_check(check_output, version):
-    check_output.return_value.decode.return_value = "patchelf {}".format(version)
+    check_output.return_value.decode.return_value = f"patchelf {version}"
     Patchelf()
 
 
 @patch("auditwheel.patcher.check_output")
 @pytest.mark.parametrize("version", ["0.8", "0.8.1", "0.1"])
 def test_patchelf_version_check_fail(check_output, version):
-    check_output.return_value.decode.return_value = "patchelf {}".format(version)
-    with pytest.raises(ValueError, match="patchelf {} found".format(version)):
+    check_output.return_value.decode.return_value = f"patchelf {version}"
+    with pytest.raises(ValueError, match=f"patchelf {version} found"):
         Patchelf()
 
 
